Add sender and reciver programs for testing xbee
The source file for the programs. Easiest way to get them to the robot is to create a couple of new programs in the mbed online compiler and replace the example program sources with these. Clicking compile on the online compiler will make a binary file that you drag and drop into the connected mbed. There should only be 1 binary file in it at a time.

Then click on the button that is on the microcontroller next to the chip. If the microcontroller is plugged in and you press the button, the lights should go off when you press the button then come back when you let go. This allows the new binary file to be used.